Created separate Character Selection script
