+++ Flonkerton-Style/Assets/scripts/CharacterSelectionScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a339512edd49a4ec299809b9653bee24
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
